+++ src/continuity/discrete/__init__.py
@@ -5,7 +5,6 @@ Functionalities handling discretization of continuous functionals.
 """
 
 from .uniform import UniformBoxSampler
+from .regular_grid import RegularGridSampler
 
-__all__ = [
-    "UniformBoxSampler",
-]
+__all__ = ["UniformBoxSampler", "RegularGridSampler"]