+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 - Move all content of `__init__.py` files to sub-modules.
 - Add `Trainer` class to replace `operator.fit` method.
 - Implement `BelNet`.
-- Add `Sampler`, `BoxSampler`, and `UniformBoxSampler` classes.
+- Add `Sampler`, `BoxSampler`, `UniformBoxSampler`, and `RegularGridSampler` classes.
 - Moved `DataLoader` into the `fit` method of the `Trainer`.
   Therefore, `Trainer.fit` expects an `OperatorDataset` now.